Ensure `storages` are cleaned up when elements are removed
All `Tracked*` collections have an internal map of 'storages'. These were not being cleaned up when elements were removed from the collection, or when the collection was cleared. That meant that every object which was ever in the collection continues to be referenced, and will never be garbage collected. Depending on the use-case, this can lead to an unbound memory leak.

This commit ensures that the storages are deleted/cleared appropriately.